+++ app/addUser/[id]/index.js
@@ -1,0 +1,8 @@
+export function POST(req) {
+	const id = req.params.id;
+	
+	return {
+		status: "User added: " + id,
+		error: false,
+	};
+}