+++ app/index.js
@@ -1,7 +1,6 @@
-export default async function Home(request) {
-	let resp = await fetch("https://jsonplaceholder.typicode.com/todos/1");
-	resp = await resp.json();
+export function GET(req) {
 	return {
-		route: resp,
+		route: "Home",
+		method: "GET",
 	};
 }