+++ app/products/index.js
@@ -1,3 +1,5 @@
 export default function Home(args) {
-	return "Products";
+	return {
+		route: "Product"
+	};
 }