+++ app/products/kitchen/[...slug]/index.js
@@ -1,5 +1,6 @@
 export default function (request) {
+	const slugs = request.params[0].split("/");
 	return {
-		about: "kitchen slugs",
+		about: slugs,
 	};
 }