+++ app/products/kitchen/index.js
@@ -1,6 +1,0 @@
-export default function Home(route) {
-	return {
-		"id": 2,
-		"type": "Kitchen"
-	};
-}