+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # folder-routing
 
+## 0.2.0
+
+### Minor Changes
+
+- Data.json, post requests and tests
+
+  All features covered here:
+
+  1. Ability to serve data from JSON files
+  2. Support of GET, POST, PUT and DELETE requests
+  3. Tests covering all requests and 4 different methods
+  4. Improved documentation
+  5. Bug Fixes
+
 ## 0.1.0
 
 ### Minor Changes